refactor(opal/time): move getSecondsUntilExpiration into @opal/time, decouple from User type
